llm-rag-chatbot/requirements.txt
streamlit
chromadb
sentence-transformers
openai
python-dotenv
numpy
pandas
scikit-learn
